Added internal API route
